backend/notification_service/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
pydantic
pydantic-settings
